+++ src/main/kotlin/com/jetbrains/snakecharm/actions/SmkGotoSymbolContributor.kt
@@ -2,7 +2,6 @@ package com.jetbrains.snakecharm.actions
 
 import com.intellij.navigation.ChooseByNameContributorEx
 import com.intellij.navigation.NavigationItem
-import com.intellij.openapi.project.Project
 import com.intellij.psi.search.GlobalSearchScope
 import com.intellij.psi.stubs.StubIndex
 import com.intellij.util.Processor