+++ src/main/kotlin/com/jetbrains/snakecharm/stringLanguage/lang/psi/references/SmkSLWildcardReference.kt
@@ -14,25 +14,26 @@ import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLExpression
 import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLReferenceExpressionImpl
 
 class SmkSLWildcardReference(
-        element: SmkSLReferenceExpressionImpl,
-        private val type: SmkWildcardsType
+    element: SmkSLReferenceExpressionImpl,
+    private val type: SmkWildcardsType,
 ) : PsiPolyVariantReferenceBase<SmkSLReferenceExpressionImpl>(element), SmkSLBaseReference, PsiReferenceEx {
 
     override fun handleElementRename(newElementName: String) =
-            element.setName(newElementName)
+        element.setName(newElementName)
 
-    override fun multiResolve(incompleteCode: Boolean): Array<out ResolveResult> =
-            type.resolveMember(
-                    wildcardName(),
-                    element,
-                    AccessDirection.READ,
-                    PyResolveContext.defaultContext()
-            ).toTypedArray()
+    override fun multiResolve(incompleteCode: Boolean): Array<out ResolveResult> {
+        val context = TypeEvalContext.codeAnalysis(element.project, element.containingFile)
+        val resolveContext = PyResolveContext.defaultContext(context)
+
+        return type.resolveMember(
+            wildcardName(), element, AccessDirection.READ, resolveContext
+        ).toTypedArray()
+    }
 
     /**
      * not canonical text!!!! (because range will be wrong for "foo.boo.doo" wildcard name
      */
-    fun wildcardName() = getWildcardTrueExpression().text
+    fun wildcardName(): String = getWildcardTrueExpression().text
 
     /**
      * In case of wildcards with dots inside, which is allows but very confusing, we need to take the
@@ -40,10 +41,10 @@ class SmkSLWildcardReference(
      * name
      */
     fun getWildcardTrueExpression() =
-            PsiTreeUtil.getTopmostParentOfType(element, SmkSLExpression::class.java) ?: element
+        PsiTreeUtil.getTopmostParentOfType(element, SmkSLExpression::class.java) ?: element
 
     override fun getVariants(): Array<out Any> =
-            type.getCompletionVariants(canonicalText, element, ProcessingContext())
+        type.getCompletionVariants(canonicalText, element, ProcessingContext())
 
     override fun getUnresolvedHighlightSeverity(context: TypeEvalContext?): HighlightSeverity? = null